Rename MULTIPLE_ITEMS to MULTIPLE_CHOICES
Translation server might come back with 200 and multiple items. This results in
COMPLETE and an array of items. This is different from 300 where only headers
are available but item to be translated still needs to be picked and translated
with a follow-up request. In order to avoid confusion, MULTIPLE_ITEMS has been
renamed to MULTIPLE_CHOICES.